+++ storybook/stories/settings/general/General.stories.js
@@ -5,12 +5,17 @@ import { boolean } from '@storybook/addon-knobs';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import SettingsWrapper from '../utils/SettingsWrapper';
-import { LANGUAGE_OPTIONS } from '../../../../source/renderer/app/config/profileConfig';
+import {
+  DATE_ENGLISH_OPTIONS,
+  LANGUAGE_OPTIONS,
+  NUMBER_OPTIONS,
+  TIME_OPTIONS,
+} from '../../../../source/renderer/app/config/profileConfig';
 import { updateParam } from '../../../addons/DaedalusMenu';
 import { locales, themesIds } from '../../_support/config';
 
 // Screens
-import GeneralSettings from '../../../../source/renderer/app/components/settings/categories/GeneralSettings';
+import ProfileSettingsForm from '../../../../source/renderer/app/components/widgets/forms/ProfileSettingsForm';
 import DisplaySettings from '../../../../source/renderer/app/components/settings/categories/DisplaySettings';
 import SupportSettings from '../../../../source/renderer/app/components/settings/categories/SupportSettings';
 import TermsOfUseSettings from '../../../../source/renderer/app/components/settings/categories/TermsOfUseSettings';
@@ -25,16 +30,21 @@ storiesOf('Settings|General', module)
   // ====== Stories ======
 
   .add('General', () => (
-    <GeneralSettings
-      languages={LANGUAGE_OPTIONS}
-      currentLocale="en-US"
-      onSelectLanguage={({ locale }) =>
-        updateParam({
-          param: 'localeName',
-          value: findKey(locales, item => item === locale),
-        })
-      }
+    <ProfileSettingsForm
       isSubmitting={boolean('isSubmitting', false)}
+      onSubmit={action('submit')}
+      onChangeItem={(param: string, value: string) => {
+        if (param === 'locale') {
+          updateParam({
+            param: 'localeName',
+            value: findKey(locales, item => item === value),
+          });
+        }
+      }}
+      currentDateFormat={DATE_ENGLISH_OPTIONS[0].value}
+      currentLocale={LANGUAGE_OPTIONS[0].value}
+      currentNumberFormat={NUMBER_OPTIONS[0].value}
+      currentTimeFormat={TIME_OPTIONS[0].value}
     />
   ))
   .add('Themes', () => (